chore: add example code to using downloader

+++ android/src/main/kotlin/com/example/pl_image_downloader/services/DownloadGlobal.kt
@@ -12,27 +12,27 @@ object DownloadGlobal {
     var downloadConfig: DownloadConfiguration = DownloadConfiguration.default()
 
     /** * A map to hold retry download tasks with their IDs as keys. */
-    var retryTask: MutableMap<Int, DownloadTask> = mutableMapOf()
+    var retryTask: MutableMap<Long, DownloadTask> = mutableMapOf()
     /** * Adds a retry task to the map if it doesn't already exist. */
-    fun addRetryTask(id: Int, task: DownloadTask) {
+    fun addRetryTask(id: Long, task: DownloadTask) {
         if(retryTask.containsKey(id)) return
         retryTask[id] = task
     }
 
     /** * Removes a retry task by its ID. */
-    fun removeRetryTask(id: Int) {
+    fun removeRetryTask(id: Long) {
         retryTask.remove(id)
     }
 
     /** * A map to hold running download tasks with their IDs as keys. */
-    var runningTask: MutableMap<Int, DownloadTask> = mutableMapOf()
-    fun addRunningTask(id: Int, task: DownloadTask) {
+    var runningTask: MutableMap<Long, DownloadTask> = mutableMapOf()
+    fun addRunningTask(id: Long, task: DownloadTask) {
         if(runningTask.containsKey(id)) return
         runningTask[id] = task
     }
 
     /** * Removes a running task by its ID. */
-    fun removeRunningTask(id: Int) {
+    fun removeRunningTask(id: Long) {
         runningTask.remove(id)
     }
